# Backend/openapi.yaml
openapi: 3.1.0
info:
  title: APIFiddle Sample Definition
  version: 0.0.0
  description: Example workspace
paths:
...
  /user/reset_password:
    post:
      summary: reset password
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/BadRequest"
      operationId: resetPassword
...
components:
  responses:
    Ok:
      content:
        application/json:
          schema:
            type: object
            required:
              - data
            properties:
              data:
                type: object
      description: The request was successful, and the server has returned the
        requested resource in the response body.
    BadRequest:
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              errors:
                type: array
                items:
                  type: object
                  required:
                    - message
                  properties:
                    message:
                      type: string
              message:
                type: string
      description: The server could not understand the request due to invalid syntax.
        The client should modify the request and try again.
...